Move `useThumbnail` into hooks.
